+++ CHANGELOG.md
@@ -13,5 +13,11 @@ From v2.1.0.a2 to v2.2.0
 ## New Features
 - Most unit tests now have a variant that checks whether using a fixed random seed gives reproducible results.
 - The build pipeline now contains a check that the jupyter notebooks give the same results as ones that were observed before.
+- Added `FitMonitor`, `AssessorMonitor`, and `HyperParameterOptimizationMonitor` base classes to monitor the progress of fitting, assessing, and  hyperparameter optimization, respectively.
+- Added `BaseMonitor` class to internally keep track of the progress of a fitting, assessing, or hyperparameter optimization process.
+- Added `FileMonitor` class to save the progress of a fitting, assessing, or hyperparameter optimization process to files.
+- Added `WandBMonitor` class to save the progress of a fitting, assessing, or hyperparameter optimization process to [Weights & Biases](https://wandb.ai/).
+- Added `NullMonitor` class to ignore the progress of a fitting, assessing, or hyperparameter optimization process.
+- Added `ListMonitor` class to combine multiple monitors.
 
 ## Removed Features